pyproject.toml
[tool.poetry]
name = "detection-models"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
tqdm = "^4.66.4"
pillow = "^11.2.1"
numpy = "^2.0.2"
tensorboard = "^2.17.0"
pyyaml = "^6.0.2"
torch = "^2.7.0"
torchvision = "^0.22.0"
scipy = "^1.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
